Better tabsystem + create preset menu + scaling
Models can now also be rotated with right mouse buttons. Players can now create new presets in the tool by choosing a name, category, model and size. When loading a prefab, the bounds of the mesh will be scaled accordingly to fit in the given size (Vector2 with x and z size). The tabsystem will now not recreate buttons every time you switch a tab, but rather create all buttons immediately and hide the buttons that are not in the selected category.